Refactored the structure of the project.
