+++ TempleOfDoom/TempleOfDoom/BusinessLogic/FactoryMethodes/DoorFactory.cs
@@ -1,7 +1,7 @@
-﻿using TempleOfDoom.DataLayer.Decorators;
+﻿using TempleOfDoom.BusinessLogic.Enums;
+using TempleOfDoom.BusinessLogic.Models.Doors;
 using TempleOfDoom.DataLayer.DTO;
 using TempleOfDoom.DataLayer.Models;
-using TempleOfDoom.Enums;
 
 public class DoorFactory
 {
