+++ TempleOfDoom/TempleOfDoom/BusinessLogic/Models/Connection.cs
