+++ TempleOfDoom/TempleOfDoom/BusinessLogic/Models/Items/Boobytrap.cs
