--- TempleOfDoom/TempleOfDoom/DataLayer/Models/Door.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using TempleOfDoom.Interfaces;
-
-namespace TempleOfDoom.DataLayer.Models
-{
-    public class Door
-    {
-        public string Type { get; set; }
-        public string Color { get; set; }
-        public int No_of_stones { get; set; }
-    }
-}
-
