+++ TempleOfDoom/TempleOfDoom/PresentationLayer/GameFrame.cs
@@ -1,5 +1,4 @@
 using TempleOfDoom.DataLayer.Models;
-using TempleOfDoom.Enums;
 using TempleOfDoom.HelperClasses;
 using TempleOfDoom.Interfaces;
