+++ TempleOfDoom/TempleOfDoom/Program.cs
@@ -3,8 +3,8 @@ using System.Text;
 using TempleOfDoom.DataLayer.DTO;
 using TempleOfDoom.DataLayer.Models;
 using TempleOfDoom.DataLayer.ReaderStrategies;
-using TempleOfDoom.DataLayer.MapperStrategies;
 using TempleOfDoom.Interfaces;
+using TempleOfDoom.BusinessLogic.MapperStrategies;
 
 namespace TempleOfDoom
 {
@@ -16,7 +16,7 @@ namespace TempleOfDoom
             Console.CursorVisible = false;
 
             // Specify the path to the level data (e.g., JSON file)
-            string levelPath = "/Users/anton/Documents/GitHub/Tempeltje/TempleOfDoom/TempleOfDoom/GameLevels/TempleOfDoom.json";
+            string levelPath = "C:\\Users\\Anton Jagers\\OneDrive\\Documenten\\Documenten\\GitHub\\TempleOfDoom\\TempleOfDoom\\TempleOfDoom\\GameLevels\\TempleOfDoom.json";
 
             // Create an instance of the JsonLevelDataReader
             ILevelDataReader levelDataReader = new JsonLevelDataReader();
